iphone should be working perfectly

+++ Resources/lib/channel_download.js
@@ -34,10 +34,10 @@ var ChannelDownload = (function(){
 		var _writeZip = function() {
 			var zipPath = Ti.Filesystem.getFile(Ti.Filesystem.applicationDataDirectory, 'channels.zip');	  
 			var extractPath = Ti.Filesystem.getFile(Ti.Filesystem.applicationDataDirectory, "channels");
-			if(extractPath.exists()) {
-				extractPath.deleteDirectory()
-				extractPath.createDirectory();
-			}
+			
+			if(extractPath.exists()) extractPath.deleteDirectory();
+			
+			extractPath.createDirectory();
 			
 		  zipPath.write(this.responseData);
 			zipfile.extract(zipPath.nativePath, extractPath.nativePath);
